all done and sprit
